Level 2 hierarchy georganiseerd, karakter prefab ingevoegd
