+++ Assets/Scenes/Main Menu.unity
@@ -591,11 +591,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855595588709, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855595588709, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855792000140, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_MinValue
@@ -603,55 +603,55 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855839057826, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855839057826, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855839057826, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMin.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855965659709, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855965659709, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958855965659709, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMin.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856483125661, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856483125661, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856483125661, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMin.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856498243488, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856498243488, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856539038425, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 0.9
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856539038425, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856712926563, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_MinValue
@@ -747,7 +747,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856852582736, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7467958856875476700, guid: 006084da93f954172bf529ccb17bb4a8, type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
