+++ Assets/Scripts/Management/ResetAllLevel2.cs
@@ -1,0 +1,15 @@
+
+
+using UnityEngine;
+
+public class ResetAllLevel2 : MonoBehaviour
+{
+    private void Update()
+    {
+        if (Input.GetKey(KeyCode.Keypad0) && Input.GetKey(KeyCode.Keypad3) && Input.GetKey(KeyCode.Keypad5))
+        {
+            Debug.Log("Reset all!");
+            PlayerPrefs.SetInt("Checkpoint1", 0);
+        }
+    }
+}
